A5C7 - huge progress
I looked at why I received such a low grade on the cipher assignment before, and saw that translation was affected by whitespace and punctuation. I took that into account and upgraded my code so that now it accounts for special characters like the ! in "hello world!" and spaces. I implemented it into the encryption for translation and substitution, will write for others now.